refactor(service): migrate imports to new model classes layout in common module
