+++ xoai-service-provider/src/main/java/io/gdcc/xoai/serviceprovider/parsers/ListRecordsParser.java
@@ -8,7 +8,7 @@
 
 package io.gdcc.xoai.serviceprovider.parsers;
 
-import io.gdcc.xoai.model.oaipmh.Record;
+import io.gdcc.xoai.model.oaipmh.results.Record;
 import io.gdcc.xoai.serviceprovider.exceptions.InvalidOAIResponse;
 import io.gdcc.xoai.serviceprovider.model.Context;
 import io.gdcc.xoai.xmlio.XmlReader;
