+++ xoai-service-provider/src/test/java/io/gdcc/xoai/serviceprovider/parsers/IdentifyParserTest.java
@@ -1,6 +1,6 @@
 package io.gdcc.xoai.serviceprovider.parsers;
 
-import io.gdcc.xoai.model.oaipmh.Identify;
+import io.gdcc.xoai.model.oaipmh.verbs.Identify;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
