+++ xoai-service-provider/src/test/java/io/gdcc/xoai/serviceprovider/parsers/ListSetsParserTest.java
@@ -8,7 +8,7 @@
 
 package io.gdcc.xoai.serviceprovider.parsers;
 
-import io.gdcc.xoai.model.oaipmh.Set;
+import io.gdcc.xoai.model.oaipmh.results.Set;
 import io.gdcc.xoai.xmlio.XmlReader;
 import io.gdcc.xoai.xmlio.exceptions.XmlReaderException;
 import org.junit.jupiter.api.Test;
